refactor(/schedule/recommend): 일정 추천 관련 클래스 패키지 이동
