+++ backend/src/main/java/kr/momo/domain/schedule/recommend/EarliestFirstSorter.java
@@ -1,19 +1,18 @@
-package kr.momo.service.schedule;
+package kr.momo.domain.schedule.recommend;
 
 import java.util.Comparator;
 import java.util.List;
-import kr.momo.domain.schedule.CandidateSchedule;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class LongTermFirstSorter implements CandidateScheduleSorter {
+public class EarliestFirstSorter implements CandidateScheduleSorter {
 
     @Override
     public void sort(List<CandidateSchedule> unorderedSchedules) {
         Comparator<CandidateSchedule> comparator = Comparator.comparing(CandidateSchedule::groupSize, Comparator.reverseOrder())
-                .thenComparing(CandidateSchedule::duration, Comparator.reverseOrder());
+                .thenComparing(CandidateSchedule::startDateTime);
         unorderedSchedules.sort(comparator);
     }
 }
