+++ backend/src/main/java/kr/momo/service/schedule/ScheduleService.java
@@ -12,10 +12,12 @@ import kr.momo.domain.availabledate.AvailableDateRepository;
 import kr.momo.domain.availabledate.AvailableDates;
 import kr.momo.domain.meeting.Meeting;
 import kr.momo.domain.meeting.MeetingRepository;
-import kr.momo.domain.schedule.CandidateSchedule;
+import kr.momo.domain.schedule.recommend.CandidateSchedule;
 import kr.momo.domain.schedule.Schedule;
 import kr.momo.domain.schedule.ScheduleBatchRepository;
 import kr.momo.domain.schedule.ScheduleRepository;
+import kr.momo.service.schedule.recommend.ScheduleRecommender;
+import kr.momo.service.schedule.recommend.ScheduleRecommenderFactory;
 import kr.momo.domain.timeslot.Timeslot;
 import kr.momo.exception.MomoException;
 import kr.momo.exception.code.AttendeeErrorCode;
