+++ backend/src/main/java/kr/momo/service/schedule/recommend/FilteredScheduleRecommender.java
@@ -1,10 +1,11 @@
-package kr.momo.service.schedule;
+package kr.momo.service.schedule.recommend;
 
 import java.util.List;
 import kr.momo.domain.attendee.AttendeeGroup;
-import kr.momo.domain.schedule.CandidateSchedule;
 import kr.momo.domain.schedule.DateAndTimeslot;
 import kr.momo.domain.schedule.ScheduleRepository;
+import kr.momo.domain.schedule.recommend.CandidateSchedule;
+import kr.momo.domain.schedule.recommend.CandidateScheduleSorterFactory;
 import org.springframework.stereotype.Component;
 
 @Component
